Add barbones Input implementation, framework app
App currently just prints to console whether you're pressing left or
right.

Also includes the framework for running the game. Implements the logic
for checking to see if the game is currently running and is in focus, as
well as properly handing off the main thread from UltraBreakoutActivity
to UltraBreakout.