Implement finer control of ball movement
When the ball hits the paddle, it will now move in the direction that it
hit the paddle in. Ex if the ball hit the left side of the paddle, it
will move with a bias towards the left side of the screen, with the
strength of this bias proportional to how "left" it hit the paddle.